requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
chromadb
pypdf
python-dotenv
python-multipart